backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "notology"
version = "0.1.0"
description = ""
authors = ["yashvisal <yashvisal@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.12"
langchain = "^0.2.11"
langchain-openai = "^0.1.20"
python-dotenv = "^1.0.1"
unstructured = {extras = ["openai", "pinecone", "s3"], version = "^0.15.0"}
langserve = {extras = ["all"], version = "^0.2.2"}
fastapi = "^0.111.1"
uvicorn = "^0.30.4"
langchain-community = "^0.2.11"
langgraph = "^0.2.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.27.0"
convex = "^0.6.0"

[tool.poetry.scripts]
start = "notology.server:start"
ingest = "notology.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"